Revert "rust: Deal with new Rust 1.66 warnings (#977)"
This reverts commit 890536c1abaef14377ebd1b8bd6fc630d3269ced.

Well, most of it, keeping only the changes in CI, but rolling back
source code to the state it was in 0.14.0.

+++ src/wrappers/themis/rust/src/secure_message.rs
@@ -183,7 +183,7 @@ impl SecureMessage {
                 return Err(error);
             }
             debug_assert!(encrypted_len <= encrypted.capacity());
-            encrypted.set_len(encrypted_len);
+            encrypted.set_len(encrypted_len as usize);
         }
 
         Ok(encrypted)
@@ -233,7 +233,7 @@ impl SecureMessage {
                 return Err(error);
             }
             debug_assert!(decrypted_len <= decrypted.capacity());
-            decrypted.set_len(decrypted_len);
+            decrypted.set_len(decrypted_len as usize);
         }
 
         Ok(decrypted)
@@ -365,7 +365,7 @@ impl SecureSign {
                 return Err(error);
             }
             debug_assert!(signed_len <= signed.capacity());
-            signed.set_len(signed_len);
+            signed.set_len(signed_len as usize);
         }
 
         Ok(signed)
@@ -493,7 +493,7 @@ impl SecureVerify {
                 return Err(error);
             }
             debug_assert!(original_len <= original.capacity());
-            original.set_len(original_len);
+            original.set_len(original_len as usize);
         }
 
         Ok(original)
